requirements/base.txt
Django>=5.1,<7.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
django-filter>=24.0
django-storages[s3]>=1.14
django-environ>=0.11
django-extensions>=3.2
psycopg[binary]>=3.1,<4.0
firebase-admin>=6.4,<7.0
celery[redis]>=5.3,<6.0
redis>=5.0,<6.0
anthropic>=0.80,<1.0
Pillow>=10.2,<13.0
qrcode[pil]>=7.4
gunicorn>=22.0
whitenoise>=6.6
drf-spectacular>=0.27
django-celery-results>=2.5
python-dateutil>=2.8
requests>=2.31
